refactor(syntax): move line terminator related variables/methods from identifier to `line_terminator` module (#16486)
Now, we have a `line_terminator` module, so moving line terminator-related variables and methods to their own module would be better for maintainability.
