+++ crates/oxc_parser/src/lexer/comment.rs
@@ -1,6 +1,6 @@
 use memchr::memmem::Finder;
 
-use oxc_syntax::identifier::is_line_terminator;
+use oxc_syntax::line_terminator::is_line_terminator;
 
 use crate::diagnostics;
